Key-error debug.
I was having a keyerror, caused by using the .drop() function to drop a column that wasn't present in my data frame, so I rectified this error by making the column first before dropping it. All this stems from the step of splitting the data from the features and target variable.